Add files via upload
The Online Banking Application is a Java-based desktop application designed to provide users with essential banking services. Built using Java and JDBC for database connectivity, the application connects to a MySQL database, ensuring data storage and retrieval for various banking operations. This application is suitable for banks aiming to offer a simple, secure, and efficient platform for managing customer accounts and transactions.